third_party/blink/web_tests/external/wpt/custom-elements/revamped-scoped-registry/Element-customElements-exceptions.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] customElements on a failed custom element created by calling createElement on CustomElementRegistry should return the registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] customElements on a failed custom element created by setting innerHTML should return the associated scoped registry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] customElements on a failed custom element created by parser should return the specified custom regsitry
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
Harness: the test ran to completion.
